Make SM100 FP8 dense path Mirage-native
