+++ include/mirage/persistent_kernel/tasks/blackwell/linear_fp8_sm100_cute_tie.cuh
@@ -4,7 +4,7 @@ namespace cute {
 
 struct ignore_t {
   template <typename T>
-  constexpr const ignore_t &operator=(T &&) const noexcept {
+  constexpr ignore_t const &operator=(T &&) const noexcept {
     return *this;
   }
 };
@@ -24,17 +24,19 @@ inline constexpr ignore_t ignore{};
 #define CUTE_TIE_OP_ASSIGN(I, TUPLE, VAR) VAR = ::cute::get<I>(TUPLE)
 
 #define CUTE_TIE_APPLY_OP_1(OP, T, V1) OP(0, T, V1);
-#define CUTE_TIE_APPLY_OP_2(OP, T, V1, V2) OP(0, T, V1); OP(1, T, V2);
+#define CUTE_TIE_APPLY_OP_2(OP, T, V1, V2)                                     \
+  OP(0, T, V1);                                                                \
+  OP(1, T, V2);
 #define CUTE_TIE_APPLY_OP_3(OP, T, V1, V2, V3)                                 \
   OP(0, T, V1);                                                                \
   OP(1, T, V2);                                                                \
   OP(2, T, V3);
-#define CUTE_TIE_APPLY_OP_4(OP, T, V1, V2, V3, V4)                            \
+#define CUTE_TIE_APPLY_OP_4(OP, T, V1, V2, V3, V4)                             \
   OP(0, T, V1);                                                                \
   OP(1, T, V2);                                                                \
   OP(2, T, V3);                                                                \
   OP(3, T, V4);
-#define CUTE_TIE_APPLY_OP_5(OP, T, V1, V2, V3, V4, V5)                        \
+#define CUTE_TIE_APPLY_OP_5(OP, T, V1, V2, V3, V4, V5)                         \
   OP(0, T, V1);                                                                \
   OP(1, T, V2);                                                                \
   OP(2, T, V3);                                                                \
@@ -42,17 +44,17 @@ inline constexpr ignore_t ignore{};
   OP(4, T, V5);
 
 #define CUTE_TIE_DECL(TUPLE_EXPR, ...)                                         \
-  auto &&CUTE_TIE_CONCAT(cute_tie__temp_tuple_, __LINE__) = (TUPLE_EXPR);     \
-  CUTE_TIE_CONCAT(CUTE_TIE_APPLY_OP_, CUTE_TIE_COUNT_ARGS(__VA_ARGS__))(      \
-      CUTE_TIE_OP_DECL,                                                        \
-      CUTE_TIE_CONCAT(cute_tie__temp_tuple_, __LINE__),                        \
-      __VA_ARGS__)
+  auto &&CUTE_TIE_CONCAT(cute_tie__temp_tuple_, __LINE__) = (TUPLE_EXPR);      \
+  CUTE_TIE_CONCAT(CUTE_TIE_APPLY_OP_, CUTE_TIE_COUNT_ARGS(__VA_ARGS__))        \
+  (CUTE_TIE_OP_DECL,                                                           \
+   CUTE_TIE_CONCAT(cute_tie__temp_tuple_, __LINE__),                           \
+   __VA_ARGS__)
 
 #define CUTE_TIE(TUPLE_EXPR, ...)                                              \
   do {                                                                         \
-    auto &&CUTE_TIE_CONCAT(cute_tie__temp_tuple_, __LINE__) = (TUPLE_EXPR);   \
-    CUTE_TIE_CONCAT(CUTE_TIE_APPLY_OP_, CUTE_TIE_COUNT_ARGS(__VA_ARGS__))(    \
-        CUTE_TIE_OP_ASSIGN,                                                    \
-        CUTE_TIE_CONCAT(cute_tie__temp_tuple_, __LINE__),                      \
-        __VA_ARGS__);                                                          \
+    auto &&CUTE_TIE_CONCAT(cute_tie__temp_tuple_, __LINE__) = (TUPLE_EXPR);    \
+    CUTE_TIE_CONCAT(CUTE_TIE_APPLY_OP_, CUTE_TIE_COUNT_ARGS(__VA_ARGS__))      \
+    (CUTE_TIE_OP_ASSIGN,                                                       \
+     CUTE_TIE_CONCAT(cute_tie__temp_tuple_, __LINE__),                         \
+     __VA_ARGS__);                                                             \
   } while (0)
