+++ include/mirage/persistent_kernel/tasks/blackwell/linear_fp8_sm100_epilogue_utils.cuh
@@ -1,14 +1,14 @@
 #pragma once
 
-#include <deep_gemm/common/types.hpp>
-#include <deep_gemm/common/utils.cuh>
+#include "linear_fp8_sm100_types.hpp"
+#include "linear_fp8_sm100_utils.cuh"
 
-namespace deep_gemm {
+namespace mirage::blackwell::linear_fp8_sm100 {
 
 struct EpilogueIdentity {
   template <uint32_t STORE_BLOCK_N>
   __device__ __forceinline__ static uint32_t
-  apply_index_n(const uint32_t &n_idx) {
+      apply_index_n(uint32_t const &n_idx) {
     return n_idx;
   }
 };
@@ -17,9 +17,8 @@ template <uint32_t kLeft, uint32_t kMid, uint32_t kRight>
 struct EpilogueHeadSplits : EpilogueIdentity {
   template <uint32_t STORE_BLOCK_N>
   __device__ __forceinline__ static uint32_t
-  apply_index_n(const uint32_t &n_idx) {
-    DG_STATIC_ASSERT(kLeft % STORE_BLOCK_N == 0 &&
-                         kMid % STORE_BLOCK_N == 0 &&
+      apply_index_n(uint32_t const &n_idx) {
+    DG_STATIC_ASSERT(kLeft % STORE_BLOCK_N == 0 && kMid % STORE_BLOCK_N == 0 &&
                          kRight % STORE_BLOCK_N == 0,
                      "Invalid head splits config");
     return n_idx + (n_idx + kRight) / (kLeft + kRight) * kMid;
@@ -28,4 +27,4 @@ struct EpilogueHeadSplits : EpilogueIdentity {
 
 #pragma clang diagnostic pop
 
-} // namespace deep_gemm
+} // namespace mirage::blackwell::linear_fp8_sm100
