+++ include/mirage/persistent_kernel/tasks/blackwell/linear_fp8_sm100_utils.cuh
@@ -4,13 +4,15 @@
 #include <cuda/std/utility>
 #include <cute/container/tuple.hpp>
 
-#include "cute_tie.cuh"
+#include "linear_fp8_sm100_cute_tie.cuh"
 
 #ifndef DG_DEVICE_ASSERT
 #define DG_DEVICE_ASSERT(cond)                                                 \
   do {                                                                         \
     if (!(cond)) {                                                             \
-      printf("Assertion failed: %s:%d, condition: %s\n", __FILE__, __LINE__,   \
+      printf("Assertion failed: %s:%d, condition: %s\n",                       \
+             __FILE__,                                                         \
+             __LINE__,                                                         \
              #cond);                                                           \
       asm("trap;");                                                            \
     }                                                                          \
@@ -29,18 +31,22 @@
 #define DG_STATIC_ASSERT(cond, ...) static_assert(cond, __VA_ARGS__)
 #endif
 
-namespace deep_gemm {
+namespace mirage::blackwell::linear_fp8_sm100 {
 
-template <typename FuncT> struct PatternVisitor {
+template <typename FuncT>
+struct PatternVisitor {
   FuncT func;
 
   __device__ __host__ explicit PatternVisitor(FuncT &&func)
       : func(std::forward<FuncT>(func)) {}
 
-  __device__ __host__ auto operator[](const uint32_t &i) { return func(i); }
+  __device__ __host__ auto operator[](uint32_t const &i) {
+    return func(i);
+  }
 };
 
-template <typename T> __device__ __host__ T ceil_div(T a, T b) {
+template <typename T>
+__device__ __host__ T ceil_div(T a, T b) {
   return (a + b - 1) / b;
 }
 
@@ -49,7 +55,8 @@ __device__ __host__ constexpr T constexpr_ceil_div(T a, T b) {
   return (a + b - 1) / b;
 }
 
-template <typename T> __device__ __host__ T align(T a, T b) {
+template <typename T>
+__device__ __host__ T align(T a, T b) {
   return ceil_div(a, b) * b;
 }
 
@@ -63,7 +70,8 @@ __device__ __host__ constexpr T constexpr_gcd(T a, T b) {
   return b == 0 ? a : constexpr_gcd(b, a % b);
 }
 
-template <typename T> __forceinline__ __device__ void swap(T &a, T &b) {
+template <typename T>
+__forceinline__ __device__ void swap(T &a, T &b) {
   T temp = a;
   a = b;
   b = temp;
@@ -81,7 +89,7 @@ __forceinline__ __device__ uint32_t get_lane_idx() {
   return lane_id;
 }
 
-__device__ __forceinline__ uint32_t ld_shared(const uint32_t *ptr) {
+__device__ __forceinline__ uint32_t ld_shared(uint32_t const *ptr) {
   uint32_t ret;
   asm volatile("ld.shared.u32 %0, [%1];"
                : "=r"(ret)
@@ -89,29 +97,25 @@ __device__ __forceinline__ uint32_t ld_shared(const uint32_t *ptr) {
   return ret;
 }
 
-__device__ __forceinline__ void st_shared(const uint32_t *ptr, uint32_t val) {
+__device__ __forceinline__ void st_shared(uint32_t const *ptr, uint32_t val) {
   asm volatile("st.shared.u32 [%0], %1;"
                :
                : "l"(__cvta_generic_to_shared(ptr)), "r"(val));
 }
 
-__device__ __forceinline__ void st_shared(const void *ptr,
-                                          uint32_t x,
-                                          uint32_t y,
-                                          uint32_t z,
-                                          uint32_t w) {
-  asm volatile("st.shared.v4.u32 [%0], {%1, %2, %3, %4};"
-               :
-               : "l"(__cvta_generic_to_shared(ptr)), "r"(x), "r"(y), "r"(z),
-                 "r"(w));
+__device__ __forceinline__ void
+    st_shared(void const *ptr, uint32_t x, uint32_t y, uint32_t z, uint32_t w) {
+  asm volatile(
+      "st.shared.v4.u32 [%0], {%1, %2, %3, %4};"
+      :
+      : "l"(__cvta_generic_to_shared(ptr)), "r"(x), "r"(y), "r"(z), "r"(w));
 }
 
 template <typename old_t>
 __device__ __forceinline__ int cast_into_bf16_and_pack(old_t &x, old_t &y) {
-  auto bf16x2 =
-      __float22bfloat162_rn({*reinterpret_cast<float *>(&x),
-                             *reinterpret_cast<float *>(&y)});
+  auto bf16x2 = __float22bfloat162_rn(
+      {*reinterpret_cast<float *>(&x), *reinterpret_cast<float *>(&y)});
   return *reinterpret_cast<int *>(&bf16x2);
 }
 
-} // namespace deep_gemm
+} // namespace mirage::blackwell::linear_fp8_sm100
