+++ include/mirage/persistent_kernel/tasks/blackwell/task_header.cuh
@@ -22,6 +22,7 @@
 // Blackwell task impls
 #include "argmax_sm100.cuh"
 #include "attention_sm100.cuh"
+#include "linear_fp8_sm100.cuh"
 #include "linear_sm100_mpk.cuh"
 #include "moe_linear_sm100.cuh"
 #include "mul_sum_add_sm100.cuh"
